requirements.txt
django>=5.0
psycopg2-binary
dj-database-url
gunicorn
pandas
openpyxl
google-generativeai